Fixes #10557: Clean-up sources tree for data source plugin
