+++ src/main/scala/bootstrap/rudder/plugin/DataSourcesConf.scala
@@ -1,26 +1,64 @@
+/*
+*************************************************************************************
+* Copyright 2016 Normation SAS
+*************************************************************************************
+*
+* This file is part of Rudder.
+*
+* Rudder is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 3 of the License, or
+* (at your option) any later version.
+*
+* In accordance with the terms of section 7 (7. Additional Terms.) of
+* the GNU General Public License version 3, the copyright holders add
+* the following Additional permissions:
+* Notwithstanding to the terms of section 5 (5. Conveying Modified Source
+* Versions) and 6 (6. Conveying Non-Source Forms.) of the GNU General
+* Public License version 3, when you create a Related Module, this
+* Related Module is not considered as a part of the work and may be
+* distributed under the license agreement of your choice.
+* A "Related Module" means a set of sources files including their
+* documentation that, without modification of the Source Code, enables
+* supplementary functions or services in addition to those offered by
+* the Software.
+*
+* Rudder is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with Rudder.  If not, see <http://www.gnu.org/licenses/>.
+
+*
+*************************************************************************************
+*/
+
 package bootstrap.rudder.plugin
 
-import com.normation.plugins.datasources.DataSourcesPluginDef
+import bootstrap.liftweb.RudderConfig
 
+import com.normation.inventory.domain.NodeId
+import com.normation.plugins.SnippetExtensionRegister
+import com.normation.plugins.datasources.DataSourceJdbcRepository
+import com.normation.plugins.datasources.DataSourceRepoImpl
+import com.normation.plugins.datasources.DataSourcesPluginDef
+import com.normation.plugins.datasources.HttpQueryDataSourceService
+import com.normation.plugins.datasources.api.DataSourceApi9
+import com.normation.plugins.datasources.api.DataSourceApiService
+import com.normation.rudder.services.policies.PromiseGenerationHooks
+import com.normation.rudder.services.servers.NewNodeManagerHooks
+import com.normation.rudder.web.rest.ApiVersion
+
+import org.joda.time.DateTime
 import org.springframework.beans.factory.InitializingBean
 import org.springframework.context.{ ApplicationContext, ApplicationContextAware }
 import org.springframework.context.annotation.{ Bean, Configuration }
-import com.normation.plugins.{ PluginName, PluginVersion, RudderPluginDef, SnippetExtensionRegister }
 
-import net.liftweb.common.Loggable
-import com.normation.inventory.domain.NodeId
-import com.normation.rudder.web.rest.datasource.DataSourceApi9
-import com.normation.rudder.services.servers.NewNodeManagerHooks
-import com.normation.rudder.datasources.DataSourceJdbcRepository
-import com.normation.rudder.datasources.DataSourceRepoImpl
-import com.normation.rudder.services.policies.PromiseGenerationHooks
-import com.normation.rudder.datasources.HttpQueryDataSourceService
-import com.normation.rudder.web.rest.datasource.DataSourceApiService
 import net.liftweb.common.Box
-import org.joda.time.DateTime
 import net.liftweb.common.Full
-import bootstrap.liftweb.RudderConfig
-import com.normation.rudder.web.rest.ApiVersion
+import net.liftweb.common.Loggable
 
 object DatasourcesConf {
